chore: Adjustment in class naming and group names
