+++ lib/Activity/Settings/FileToSign.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 /**
- * SPDX-FileCopyrightText: 2020-2025 LibreCode coop and contributors
+ * SPDX-FileCopyrightText: 2025 LibreCode coop and contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
@@ -10,7 +10,7 @@ namespace OCA\Libresign\Activity\Settings;
 
 use OCP\IL10N;
 
-class FileToSign extends SignActivitySettings {
+class FileToSign extends LibresignActivitySettings {
 	public function __construct(
 		protected IL10N $l,
 	) {
@@ -30,36 +30,5 @@ class FileToSign extends SignActivitySettings {
 	 */
 	public function getName(): string {
 		return $this->l->t('You have a <strong>file to sign</strong>');
-	}
-
-	/**
-	 * {@inheritdoc}
-	 */
-	public function getPriority(): int {
-		return 51;
-	}
-	/**
-	 * {@inheritdoc}
-	 */
-	public function canChangeNotification(): bool {
-		return true;
-	}
-	/**
-	 * {@inheritdoc}
-	 */
-	public function canChangeMail() {
-		return true;
-	}
-	/**
-	 * {@inheritdoc}
-	 */
-	public function isDefaultEnabledMail() {
-		return true;
-	}
-	/**
-	 * {@inheritdoc}
-	 */
-	public function isDefaultEnabledNotification(): bool {
-		return true;
 	}
 }
